Tried reading in JHU data
Attempted to figure out how to use JHU data. May need to import additional libraries such as selenium and install a webdriver to read the full html file. This could be that JHU usees JavaScript to load additional data after the html and why we can't see it with just the requests library.